requirements.dev.txt
python-dotenv==0.15.0
aiohttp~=3.7.2
ipython==7.19.0